troquei o flatlist das temporadas da página home para as transferencias(mudar nome para temporadas) e adicionei um texto na home
