--- app/src/screens/appScreens/Lendas/Lendas.js
@@ -1,12 +1,0 @@
-import React from 'react'
-import { StyleSheet, Text, View } from 'react-native'
-
-export default function Lendas() {
-    return (
-        <View>
-            <Text>Lendas</Text>
-        </View>
-    )
-}
-
-const styles = StyleSheet.create({})
